Changed type of file size in FileSystem.updateFile() from Int to Long.
The type Int was wrong, as there is no enforced size restriction for
the files that can be uploaded.

Changing the type did not cause any problems with existing
implementations; for the file representations already the type Long
was used.

+++ core/src/main/scala/com/github/cloudfiles/core/FileSystem.scala
@@ -191,7 +191,7 @@ trait FileSystem[ID, FILE, FOLDER, FOLDER_CONTENT] {
    * @param system the actor system
    * @return the ''Operation'' to upload new file content
    */
-  def updateFileContent(fileID: ID, size: Int, content: Source[ByteString, Any])
+  def updateFileContent(fileID: ID, size: Long, content: Source[ByteString, Any])
                        (implicit system: ActorSystem[_]): Operation[Unit]
 
   /**
